factor out the inline code
