+++ lib/ast.js
@@ -2869,6 +2869,21 @@ function walk(node, cb, to_visit = [node]) {
     return false;
 }
 
+/**
+ * Walks an AST node and its children.
+ *
+ * {cb} can return `walk_abort` to interrupt the walk.
+ *
+ * @param node
+ * @param cb {(node, info: { parent: (nth) => any }) => (boolean | undefined)}
+ *
+ * @returns {boolean} whether the walk was aborted
+ *
+ * @example
+ * const found_some_cond = walk_parent(my_ast_node, (node, { parent }) => {
+ *   if (some_cond(node, parent())) return walk_abort
+ * });
+ */
 function walk_parent(node, cb, initial_stack) {
     const to_visit = [node];
     const push = to_visit.push.bind(to_visit);
@@ -2984,6 +2999,15 @@ class TreeWalker {
         for (var i = stack.length; --i >= 0;) {
             var x = stack[i];
             if (x instanceof type) return x;
+        }
+    }
+
+    find_scope() {
+        for (let i = 0;;i++) {
+            const p = this.parent(i);
+            if (p instanceof AST_Toplevel) return p;
+            if (p instanceof AST_Lambda) return p;
+            if (p.block_scope) return p.block_scope;
         }
     }
